# services/fire-viewer-backend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fire-Viewer API",
    "description": "Transactional incident registry with idempotent ingestion, conservative spatial matching, immutable audit events and viewer manifests.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/incident/detect": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Detect Incident",
        "operationId": "detect_incident_api_v1_incident_detect_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "X-Source-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 32,
                  "maxLength": 256
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Source-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionResponse"
                }
              }
            }
          },
          "201": {
            "description": "A new incident series was created."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incident/{fire_id}": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Get Incident",
        "operationId": "get_incident_api_v1_incident__fire_id__get",
        "parameters": [
          {
            "name": "fire_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fire Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentPublicResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incident/{fire_id}/public-view": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Get Public View",
        "operationId": "get_public_view_api_v1_incident__fire_id__public_view_get",
        "parameters": [
          {
            "name": "fire_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fire Id"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicIncidentView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incident/{fire_id}/public-view/export.json": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Export Public View Json",
        "operationId": "export_public_view_json_api_v1_incident__fire_id__public_view_export_json_get",
        "parameters": [
          {
            "name": "fire_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fire Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incident/{fire_id}/public-view/timeline.csv": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Export Public Timeline Csv",
        "operationId": "export_public_timeline_csv_api_v1_incident__fire_id__public_view_timeline_csv_get",
        "parameters": [
          {
            "name": "fire_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fire Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incident/{fire_id}/reports": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Create Public Report",
        "operationId": "create_public_report_api_v1_incident__fire_id__reports_post",
        "parameters": [
          {
            "name": "fire_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fire Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicIncidentReportRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicIncidentReportReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incident/{fire_id}/manifest": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Get the public viewer manifest",
        "description": "Returns the public ViewerManifest in snake_case. A matching If-None-Match returns 304 without a body. This read endpoint never returns 409; conflicts are reserved for mutations.",
        "operationId": "get_manifest_api_v1_incident__fire_id__manifest_get",
        "parameters": [
          {
            "name": "fire_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fire Id"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current public ViewerManifest.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewerManifest"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "Strong entity tag calculated from the serialized ViewerManifest.",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "description": "Short public cache lifetime for this viewer representation.",
                "schema": {
                  "type": "string",
                  "example": "public, max-age=30, must-revalidate"
                }
              },
              "X-Trace-Id": {
                "description": "Trace identifier to include when reporting an error.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Manifest unchanged; the response body is empty.",
            "headers": {
              "ETag": {
                "description": "Strong entity tag calculated from the serialized ViewerManifest.",
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "description": "Short public cache lifetime for this viewer representation.",
                "schema": {
                  "type": "string",
                  "example": "public, max-age=30, must-revalidate"
                }
              },
              "X-Trace-Id": {
                "description": "Trace identifier to include when reporting an error.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "The fire_id path parameter has an invalid format.",
            "headers": {
              "X-Trace-Id": {
                "description": "Trace identifier to include when reporting an error.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "trace_id"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "trace_id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No incident exists for this fire_id.",
            "headers": {
              "X-Trace-Id": {
                "description": "Trace identifier to include when reporting an error.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "trace_id"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "trace_id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "410": {
            "description": "The incident has been tombstoned.",
            "headers": {
              "X-Trace-Id": {
                "description": "Trace identifier to include when reporting an error.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "trace_id"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "trace_id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "The incident data is temporarily unavailable.",
            "headers": {
              "X-Trace-Id": {
                "description": "Trace identifier to include when reporting an error.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "trace_id"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "trace_id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/incidents/{fire_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Incident",
        "operationId": "get_incident_api_v1_admin_incidents__fire_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fire_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fire Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminIncidentDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DetectionRequest": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/SourceInput"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          },
          "geometry": {
            "$ref": "#/components/schemas/PointGeometryInput"
          },
          "evidence": {
            "$ref": "#/components/schemas/EvidenceInput"
          },
          "context": {
            "$ref": "#/components/schemas/DetectionContext"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "observed_at",
          "received_at",
          "geometry",
          "evidence",
          "context"
        ],
        "title": "DetectionRequest"
      },
      "DetectionResponse": {
        "properties": {
          "observation_id": {
            "type": "string",
            "title": "Observation Id"
          },
          "decision": {
            "$ref": "#/components/schemas/MatchDecision"
          },
          "fire_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fire Id"
          },
          "episode_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Id"
          },
          "proposed_fire_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Fire Id"
          },
          "proposed_episode_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Episode Id"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "margin_to_second_candidate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Margin To Second Candidate"
          },
          "factors": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Factors"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "review_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Review Reasons"
          },
          "public_confirmation": {
            "type": "string",
            "enum": [
              "pending",
              "corroborated",
              "not_applicable"
            ],
            "title": "Public Confirmation"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "observation_id",
          "decision",
          "public_confirmation",
          "policy_id",
          "trace_id"
        ],
        "title": "DetectionResponse"
      },
...
      "EvidenceInput": {
        "properties": {
          "content_hash": {
            "type": "string",
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Content Hash"
          },
          "license": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "License"
          },
          "external_reference": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "External Reference"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "content_hash",
          "license"
        ],
        "title": "EvidenceInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IncidentPublicResponse": {
        "properties": {
          "fire_id": {
            "type": "string",
            "title": "Fire Id"
          },
          "canonical_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Name"
          },
          "visibility": {
            "$ref": "#/components/schemas/PublicVisibility"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "current_episode_id": {
            "type": "string",
            "title": "Current Episode Id"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PointGeometryInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "public_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Note"
          },
          "last_observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Observed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "episodes": {
            "items": {
              "$ref": "#/components/schemas/EpisodeSummary"
            },
            "type": "array",
            "title": "Episodes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fire_id",
          "visibility",
          "status",
          "current_episode_id",
          "location",
          "last_observed_at",
          "created_at",
          "episodes"
        ],
        "title": "IncidentPublicResponse"
      },
...
      "MatchDecision": {
        "type": "string",
        "enum": [
          "create",
          "attach",
          "review"
        ],
        "title": "MatchDecision"
      },
...
      "PublicIncidentReportReceipt": {
        "properties": {
          "receipt_id": {
            "type": "string",
            "title": "Receipt Id"
          },
          "status": {
            "type": "string",
            "const": "received",
            "title": "Status",
            "default": "received"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "replayed": {
            "type": "boolean",
            "title": "Replayed",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "receipt_id",
          "submitted_at"
        ],
        "title": "PublicIncidentReportReceipt"
      },
      "PublicIncidentReportRequest": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/PublicReportCategory"
          },
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 12,
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "category",
          "message"
        ],
        "title": "PublicIncidentReportRequest"
      },
...
      "SourceInput": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 3,
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/SourceType"
          },
          "trust": {
            "$ref": "#/components/schemas/SourceTrust",
            "default": "unverified"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type"
        ],
        "title": "SourceInput"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}